CHORE: QOL: isolated changes to batched_helper.py
